app/requirements.txt
Django>=5.1,<5.2
psycopg2-binary>=2.9
PyYAML>=6.0
gunicorn>=22.0
openai>=1.30
